add validation leave type and custom response

+++ HR.LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
+﻿using HR.LeaveManagement.Application.Exceptions;
+using HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
 using MediatR;
 using System.Threading;
@@ -16,6 +17,9 @@ namespace HR.LeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leavetype = await _leaveTypeRepository.Get(request.Id);
+            if (leavetype == null)
+                throw new NotFoundException(nameof(leavetype), request.Id);
+
             await _leaveTypeRepository.Delete(leavetype);
 
             return Unit.Value;
